# [ADP-3272] Simplify creation of preselected UTxO index.