--- a/lib/balance-tx/lib/internal/Internal/Cardano/Write/Tx/Balance.hs
+++ b/lib/balance-tx/lib/internal/Internal/Cardano/Write/Tx/Balance.hs
@@ -592,8 +592,6 @@ balanceTransaction
             $ fmap convertUTxO
             $ Map.toList
             $ unUTxO selectedUTxO
-        txIns :: [TxIn]
-        txIns = Set.toList txInSet
         txInSet :: Set TxIn
         txInSet = tx ^. bodyTxL . inputsTxBodyL